requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
scikit-learn>=1.3.0
plotly>=5.18.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
pymysql>=1.1.0
cryptography>=41.0.0
altair>=5.0.0